+++ example/lib/main.dart
@@ -125,6 +125,7 @@ class _HomePageState extends State<HomePage> {
                     },
                   ),
                   16.height,
+                  16.height,
                   TextButton(
                     onPressed: () {
                       AnimatedListViewExample().launch(context);