+++ lib/src/extensions/string_extensions.dart
@@ -302,4 +302,12 @@ extension StringExtension on String? {
 
     return caseSearchList;
   }
+
+  /// Returns true if given value is '1', else returns false
+  bool getBoolInt() {
+    if (this == "1") {
+      return true;
+    }
+    return false;
+  }
 }